+++ changelog.md
@@ -13,6 +13,7 @@
 - React debug errors about missing "key" prop.
 - Corrected "Requires at least" constraint to WordPress version `6.6.0`.
 - Resolved "wp.editPost.PluginDocumentSettingPanel is deprecated since version 6.6" console warning in the block editor.
+- Asana tag options would not load in Settings > Workspace when using the "Plain" permalink structure.
 
 ### 4.6.0 - 2025-01-26